Fix crash on new client, update node-osc, relay oscServer to oscClient
Crash was caused by trying to rebind to the server socket -- avoided
by caching the servers for a port. Update to node-osc wasn't strictly
needed, but was done to prepare for forking it to support relaying
oscServer to oscClient -- this will allow the bridge to send and
receive on the same udp socket enabling proper x32 functionality.